Extract n-body code into library and make binary into demo example
This allows reuse of the implementation of the case study.
