+++ .idea/runConfigurations/Run_n_body.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Run n-body" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
-    <option name="command" value="run --package n-body --bin n-body --release" />
+    <option name="command" value="run --package n-body --example n_body_demo --release" />
     <option name="workingDirectory" value="file://$PROJECT_DIR$" />
     <option name="emulateTerminal" value="false" />
     <option name="channel" value="DEFAULT" />
