+++ .idea/runConfigurations/Run_n_body__with_profiling_.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Run n-body (with profiling)" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
-    <option name="command" value="run --package n-body --bin n-body --release --features profile" />
+    <option name="command" value="run --package n-body --example n_body_demo --release --features profile" />
     <option name="workingDirectory" value="file://$PROJECT_DIR$" />
     <option name="emulateTerminal" value="false" />
     <option name="channel" value="DEFAULT" />
