+++ crates/n-body/src/lib.rs
@@ -1,69 +1,28 @@
-#[allow(unused)] // When swapping between scenes some will always go unused.
-mod scenes;
+pub mod scenes;
 
-use crate::scenes::*;
-use cgmath::{Deg, InnerSpace, MetricSpace, Point3, Vector3, Zero};
+use cgmath::{InnerSpace, MetricSpace, Point3, Vector3, Zero};
 use color_eyre::Report;
-use crossbeam::channel::unbounded;
 use ecs::systems::{Query, Read, Write};
-use ecs::{Application, ApplicationBuilder, BasicApplicationBuilder};
+use ecs::Application;
 use gfx_plugin::rendering::Position;
-use gfx_plugin::Graphical;
 use rand::distributions::Uniform;
 use rand::prelude::Distribution;
 use rand::Rng;
-use scheduler::executor::WorkerPool;
-use scheduler::schedule::PrecedenceGraph;
 use tracing::instrument;
-
-#[instrument]
-fn main() -> GenericResult<()> {
-    let mut app_builder = BasicApplicationBuilder::default().with_rendering()?;
-
-    #[cfg(feature = "profile")]
-    {
-        use ecs::profiling::Profileable;
-        app_builder = app_builder.with_profiling()?;
-    }
-    #[cfg(not(feature = "profile"))]
-    {
-        use ecs::logging::Loggable;
-        app_builder = app_builder.with_tracing()?;
-    }
-
-    let mut app = app_builder
-        .output_directory(env!("OUT_DIR"))
-        .light_model("sphere.obj")
-        .field_of_view(Deg(90.0))
-        .far_clipping_plane(10_000.0)
-        .camera_movement_speed(100.0)
-        .add_system(movement)
-        .add_system(acceleration)
-        .add_system(gravity)
-        .build()?;
-
-    SMALL_HEAVY_CLUSTERS.spawn_bodies(&mut app, create_rendered_planet_entity)?;
-    SINGLE_HEAVY_BODY_AT_ORIGIN.spawn_bodies(&mut app, create_rendered_sun_entity)?;
-
-    let (_shutdown_sender, shutdown_receiver) = unbounded();
-    app.run::<WorkerPool, PrecedenceGraph>(shutdown_receiver)?;
-
-    Ok(())
-}
 
 // todo(#90): change to use dynamic delta time.
 // todo(#90) currently assuming a hardcoded tick rate.
 const FIXED_TIME_STEP: f32 = 1.0 / 20000.0;
 
-trait BodySpawner {
+pub trait BodySpawner {
     fn spawn_bodies<App, CreateEntityFn>(
         &self,
         app: &mut App,
         create_entity: CreateEntityFn,
     ) -> GenericResult<()>
-    where
-        App: Application,
-        CreateEntityFn: Fn(&mut App, Position, Mass, Velocity, Acceleration) -> GenericResult<()>;
+        where
+            App: Application,
+            CreateEntityFn: Fn(&mut App, Position, Mass, Velocity, Acceleration) -> GenericResult<()>;
 }
 
 // Need a wrapper because the trait can't be implemented on foreign types.
@@ -83,7 +42,7 @@ impl Distribution<RandomPosition> for Uniform<f32> {
 
 /// The mass (in kilograms) of a body.
 #[derive(Debug)]
-struct Mass(f64);
+pub struct Mass(f64);
 
 impl Distribution<Mass> for Uniform<f64> {
     fn sample<R: Rng + ?Sized>(&self, random: &mut R) -> Mass {
@@ -93,7 +52,7 @@ impl Distribution<Mass> for Uniform<f64> {
 
 /// How fast a body is moving, in meters/second.
 #[derive(Debug)]
-struct Velocity(Vector3<f64>);
+pub struct Velocity(Vector3<f64>);
 
 impl Distribution<Velocity> for Uniform<f64> {
     fn sample<R: Rng + ?Sized>(&self, random: &mut R) -> Velocity {
@@ -110,7 +69,7 @@ impl Distribution<Velocity> for Uniform<f64> {
 
 /// How fast a body is accelerating, in meters/second^2.
 #[derive(Debug)]
-struct Acceleration(Vector3<f64>);
+pub struct Acceleration(Vector3<f64>);
 
 impl Distribution<Acceleration> for Uniform<f64> {
     fn sample<R: Rng + ?Sized>(&self, random: &mut R) -> Acceleration {
@@ -125,11 +84,11 @@ impl Distribution<Acceleration> for Uniform<f64> {
     }
 }
 
-type GenericResult<T> = Result<T, Report>;
+pub type GenericResult<T> = Result<T, Report>;
 
 #[instrument(skip_all)]
 #[cfg_attr(feature = "profile", inline(never))]
-fn movement(mut position: Write<Position>, velocity: Read<Velocity>) {
+pub fn movement(mut position: Write<Position>, velocity: Read<Velocity>) {
     let Velocity(velocity) = *velocity;
 
     position.point += velocity.map(|coord| coord as f32) * FIXED_TIME_STEP;
@@ -137,7 +96,7 @@ fn movement(mut position: Write<Position>, velocity: Read<Velocity>) {
 
 #[instrument(skip_all)]
 #[cfg_attr(feature = "profile", inline(never))]
-fn acceleration(mut velocity: Write<Velocity>, acceleration: Read<Acceleration>) {
+pub fn acceleration(mut velocity: Write<Velocity>, acceleration: Read<Acceleration>) {
     let Velocity(ref mut velocity) = *velocity;
     let Acceleration(acceleration) = *acceleration;
 
@@ -146,7 +105,7 @@ fn acceleration(mut velocity: Write<Velocity>, acceleration: Read<Acceleration>)
 
 #[instrument(skip_all)]
 #[cfg_attr(feature = "profile", inline(never))]
-fn gravity(
+pub fn gravity(
     position: Read<Position>,
     mut acceleration: Write<Acceleration>,
     bodies_query: Query<(Read<Position>, Read<Mass>)>,
